+++ Backend/pratik-dairy-product/product-service/src/main/java/com/pratikdairy/product/repository/ProductRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -17,14 +18,15 @@ public interface ProductRepository extends JpaRepository<Product, String> {
     // Atomic, race-condition-safe stock decrement.
     // Only succeeds (returns 1) if there is enough stock at the moment of the UPDATE.
     // This is what prevents two simultaneous orders from overselling the same product.
+    // `quantity` is weight-adjusted (stockUnit-multiples), e.g. 0.5 for "250g" x2 off a "1kg" stockUnit.
     @Modifying
     @Query("UPDATE Product p SET p.stockQuantity = p.stockQuantity - :quantity " +
             "WHERE p.id = :id AND p.stockQuantity >= :quantity")
-    int decrementStock(@Param("id") String id, @Param("quantity") int quantity);
+    int decrementStock(@Param("id") String id, @Param("quantity") BigDecimal quantity);
 
     // Used to roll back a decrement if a later item in the same order fails stock validation.
     @Modifying
     @Query("UPDATE Product p SET p.stockQuantity = p.stockQuantity + :quantity WHERE p.id = :id")
-    int restoreStock(@Param("id") String id, @Param("quantity") int quantity);
+    int restoreStock(@Param("id") String id, @Param("quantity") BigDecimal quantity);
 
 }
